+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added `table_columns` parameter to `civis.io.civis_file_to_table`, `civis.io.dataframe_to_civis`, and `civis.io.csv_to_civis` (#379)
 
 ### Fixed
-
+- No longer require ServiceClient to be instantiated to parse a
+  service api spec. (#382)
 - Fixed/relaxed version specifications for click, jsonref, and jsonschema. (#377)
 
 ### Removed